+++ app/teacher/dashboard/page.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { Header } from "@/components/ui/adminheader"
+import { Header } from "@/components/ui/teacherheader"
 import { CiCirclePlus } from "react-icons/ci";
 import { useState } from "react";
 import {Checkbox } from "../../dashboard/ui/checkbox";
